+++ packages/opentelemetry-node/CHANGELOG.md
@@ -16,6 +16,9 @@
   OpenTelemetry JS APIs that are **not yet stable**. These APIs may break in
   *minor* versions of `@elastic/opentelemetry-node`.
 
+- chore: add custom GCP resource detector to avoid internal telemetry to be sent. See <https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2320>.
+  (https://github.com/elastic/elastic-otel-node/pull/709)
+
 ## v1.0.0
 
 - BREAKING CHANGE: Change the default behavior of logging framework